+++ ftplugin/markdown.lua
@@ -1,1 +1,8 @@
 vim.b.slime_cell_delimiter = "```"
+
+-- wrap text, but by word no character
+-- indent the wrappped line
+vim.wo.wrap = true
+vim.wo.linebreak = true
+vim.wo.breakindent = true
+vim.wo.showbreak = '|'