+++ ftplugin/neotest-output.lua
@@ -1,0 +1,2 @@
+vim.keymap.set('n', 'q', ':close<cr>', {silent=true, buffer=0})
+