+++ init.lua
@@ -2,4 +2,7 @@ require 'config.global'
 require 'config.lazy'
 require 'config.autocommands'
 require 'config.keymap'
+require 'config.color'
+
+require'config.experimental'