+++ lua/config/global.lua
@@ -1,13 +1,16 @@
-vim.g.markdown_fenced_languages = { 'html', 'python', 'bash=sh', 'R=r' }
+
+-- proper colors
 vim.opt.termguicolors = true
 
 -- more opinionated
 vim.opt.number = true                  -- show linenumbers
-vim.opt.timeoutlen = 400               -- until which-key pops up
-vim.opt.updatetime = 250               -- for autocommands and hovers
 vim.opt.mouse = 'a'                    -- enable mouse
 vim.opt.mousefocus = true
 vim.opt.clipboard:append 'unnamedplus' -- use system clipboard
+
+ 
+vim.opt.timeoutlen = 400               -- until which-key pops up
+vim.opt.updatetime = 250               -- for autocommands and hovers
 
 -- don't ask about existing swap files
 vim.opt.shortmess:append 'A'
@@ -25,9 +28,6 @@ vim.g.maplocalleader = " "
 -- smarter search
 vim.opt.ignorecase = true
 vim.opt.smartcase = true
-
--- colortheme tweaks
-vim.g.nord_borders = true
 
 -- indent
 vim.opt.smartindent = true
@@ -48,6 +48,21 @@ vim.opt.foldlevel = 99
 -- global statusline
 vim.opt.laststatus = 3
 
+vim.cmd [[
+let g:currentmode={
+       \ 'n'  : '%#String# NORMAL ',
+       \ 'v'  : '%#Search# VISUAL ',
+       \ "\<C-V>" : '%#Title# V·Block ',
+       \ 'V'  : '%#IncSearch# V·Line ',
+       \ 'Rv' : '%#String# V·Replace ',
+       \ 'i'  : '%#ModeMsg# INSERT ',
+       \ 'R'  : '%#Substitute# R ',
+       \ 'c'  : '%#CurSearch# Command ',
+       \ 't'  : '%#ModeMsg# TERM ',
+       \}
+]]
+vim.opt.statusline = '%{%g:currentmode[mode()]%} %* %t | %y | %* %= c:%c l:%l/%L %p%% 🦦 '
+
 -- split right and below by default
 vim.opt.splitright = true
 vim.opt.splitbelow = true
@@ -56,10 +71,14 @@ vim.opt.splitbelow = true
 vim.opt.showtabline = 1
 
 --windowline
-vim.opt.winbar = '%t'
+vim.opt.winbar = '%f'
 
 --don't continue comments automagically
 vim.opt.formatoptions:remove({ 'c', 'r', 'o' })
 
 -- hide cmdline when not used
 vim.opt.cmdheight = 0
+
+-- scroll before end of window
+vim.opt.scrolloff = 5
+