+++ lua/plugins/specific.lua
@@ -1,0 +1,14 @@
+-- plugins for specifc usecases that
+-- not everyone will need
+
+return {
+
+  -- {
+  --   "jakewvincent/mkdnflow.nvim",
+  --   config = function ()
+  --     local mkdnflow = require("mkdnflow")
+  --     mkdnflow.setup{}
+  --   end
+  -- }
+
+}