+++ lua/plugins/ui.lua
@@ -56,86 +56,74 @@ return {
       telescope.load_extension('ui-select')
       telescope.load_extension('file_browser')
       telescope.load_extension('dap')
-      telescope.load_extension('project')
     end
   },
   { 'nvim-telescope/telescope-ui-select.nvim' },
   { 'nvim-telescope/telescope-fzf-native.nvim',  build = 'make' },
   { 'nvim-telescope/telescope-dap.nvim' },
   { 'nvim-telescope/telescope-file-browser.nvim' },
-  { 'nvim-telescope/telescope-project.nvim' },
+
   {
     'nvim-lualine/lualine.nvim',
-    dependencies = {
-      -- { 'f-person/git-blame.nvim' },
-    },
     config = function()
-      -- local git_blame = require('gitblame')
-      -- vim.g.gitblame_display_virtual_text = 0
-      vim.o.shortmess = vim.o.shortmess ..
-      "S"                                      -- this is for the search_count function so lua can know this is `lua expression`
-      --function for optimizing the search count
-      local function search_count()
-        if vim.api.nvim_get_vvar("hlsearch") == 1 then
-          local res = vim.fn.searchcount({ maxcount = 999, timeout = 500 })
 
-          if res.total > 0 then
-            return string.format("%d/%d", res.current, res.total)
-          end
+      local function macro_recording()
+        local reg = vim.fn.reg_recording()
+        if reg == '' then
+          return ''
         end
+        return '📷[' .. reg .. ']'
+      end
 
-        return ""
-      end
-      local function macro_reg()
-        return vim.fn.reg_recording()
-      end
       require('lualine').setup {
         options = {
           section_separators = '',
           component_separators = '',
           globalstatus = true,
-          theme = "catppuccin",
         },
         sections = {
-          lualine_a = { 'mode', { macro_reg, type = 'lua_expr', color = 'WarningMsg' } },
-          lualine_b = { 'branch', { search_count, type = 'lua_expr' } },
-          lualine_c = {
-            -- { git_blame.get_current_blame_text, cond = git_blame.is_blame_text_available }
-          },
+          lualine_a = { 'mode', macro_recording },
+          lualine_b = { 'branch', 'diff', 'diagnostics' },
+          -- lualine_b = {},
+          lualine_c = { 'searchcount' },
+          lualine_x = { 'filetype' },
+          lualine_y = { 'progress' },
+          lualine_z = { 'location' }
         },
         extensions = { 'nvim-tree' },
       }
     end
   },
+
   {
     'nanozuki/tabby.nvim',
     config = function()
       require 'tabby.tabline'.use_preset('tab_only')
     end
   },
-  -- {
-  --   'dstein64/nvim-scrollview',
-  --   config = function()
-  --     require('scrollview').setup({
-  --       current_only = true,
-  --       signs_on_startup = {}
-  --     })
-  --   end
-  -- },
+
+  {
+    'dstein64/nvim-scrollview',
+    config = function()
+      require('scrollview').setup({
+        current_only = true,
+      })
+    end
+  },
+
   -- { 'RRethy/vim-illuminate' }, -- highlight current word
+
   -- filetree
-  { 'nvim-tree/nvim-tree.lua',
+  {
+    'nvim-tree/nvim-tree.lua',
     keys = {
-      { '<c-b>', ':NvimTreeToggle<cr>' },
+      { '<c-b>', ':NvimTreeToggle<cr>', desc='toggle nvim-tree' },
     },
     config = function()
       require 'nvim-tree'.setup {
         disable_netrw       = true,
         update_focused_file = {
           enable = true,
-        },
-        filesystem_watchers  = {
-          enable = false,
         },
         git                 = {
           enable = true,
@@ -150,12 +138,18 @@ return {
   },
   -- show keybinding help window
   { 'folke/which-key.nvim' },
+
   {
     'simrat39/symbols-outline.nvim',
+    cmd = 'SymbolsOutline',
+    keys = {
+      { '<leader>lo', ':SymbolsOutline<cr>', desc = 'symbols outline' },
+    },
     config = function()
       require("symbols-outline").setup()
     end
   },
+
   -- terminal
   {
     "akinsho/toggleterm.nvim",
@@ -174,13 +168,34 @@ return {
       require("trouble").setup {}
     end
   },
+
   {
     'lukas-reineke/indent-blankline.nvim',
     config = function()
-      require("indent_blankline").setup {
+      require("ibl").setup {
         show_current_context = true,
         show_current_context_start = false,
       }
     end
   },
+
+  {
+    'lukas-reineke/headlines.nvim',
+    dependencies = "nvim-treesitter/nvim-treesitter",
+    config = function()
+      require("headlines").setup {
+        quarto = {
+          query = vim.treesitter.query.parse(
+            "markdown",
+            [[
+                (fenced_code_block) @codeblock
+            ]]),
+          codeblock_highlight = "CodeBlock",
+          treesitter_language = "markdown",
+        },
+      }
+    end
+  }
+
+
 }